perf(datagrid): split Excel/PDF export into lazy-loadable Lumeo.DataGrid.Export
ClosedXML (Excel) and QuestPDF (PDF) were referenced by the core Lumeo package,
so their ~1.6 MB of dependencies (DocumentFormat.OpenXml, SixLabors.Fonts, …)
loaded eagerly on every first visit even though only the DataGrid export action
uses them. Because Lumeo.dll is always loaded first, the only way to defer them
is to remove them from the core's dependency closure.

Move the heavy Excel/PDF implementation into a new Lumeo.DataGrid.Export assembly
that registers itself with the core via a module initializer. The core keeps the
IDataGridExportService facade (CSV + download stay dependency-free) and delegates
Excel/PDF to the backend, resolving it on demand. Lumeo.DataGrid ships the new
package transitively, so export works out of the box with no API change.

The docs app marks the export assemblies BlazorWebAssemblyLazyLoad and wires
DataGridExportLoader so the DataGrid fetches them only on the first export click.
QuestPDF (PDF) is loaded only on a PDF export, so Excel-only WASM users never
download it (PDF export throws on browser-wasm anyway).

Bumps version to 3.7.0.
