+++ .github/workflows/publish.yml
@@ -49,6 +49,7 @@ jobs:
           dotnet restore src/Lumeo.Charts/Lumeo.Charts.csproj
           dotnet restore src/Lumeo.CodeEditor/Lumeo.CodeEditor.csproj
           dotnet restore src/Lumeo.DataGrid/Lumeo.DataGrid.csproj
+          dotnet restore src/Lumeo.DataGrid.Export/Lumeo.DataGrid.Export.csproj
           dotnet restore src/Lumeo.Editor/Lumeo.Editor.csproj
           dotnet restore src/Lumeo.FileViewer/Lumeo.FileViewer.csproj
           dotnet restore src/Lumeo.Scheduler/Lumeo.Scheduler.csproj
@@ -65,6 +66,7 @@ jobs:
           dotnet build src/Lumeo.Charts/Lumeo.Charts.csproj -c Release --no-restore -warnaserror
           dotnet build src/Lumeo.CodeEditor/Lumeo.CodeEditor.csproj -c Release --no-restore -warnaserror
           dotnet build src/Lumeo.DataGrid/Lumeo.DataGrid.csproj -c Release --no-restore -warnaserror
+          dotnet build src/Lumeo.DataGrid.Export/Lumeo.DataGrid.Export.csproj -c Release --no-restore -warnaserror
           dotnet build src/Lumeo.Editor/Lumeo.Editor.csproj -c Release --no-restore -warnaserror
           dotnet build src/Lumeo.FileViewer/Lumeo.FileViewer.csproj -c Release --no-restore -warnaserror
           dotnet build src/Lumeo.Scheduler/Lumeo.Scheduler.csproj -c Release --no-restore -warnaserror
@@ -86,6 +88,7 @@ jobs:
           dotnet pack src/Lumeo.Charts/Lumeo.Charts.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
           dotnet pack src/Lumeo.CodeEditor/Lumeo.CodeEditor.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
           dotnet pack src/Lumeo.DataGrid/Lumeo.DataGrid.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
+          dotnet pack src/Lumeo.DataGrid.Export/Lumeo.DataGrid.Export.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
           dotnet pack src/Lumeo.Editor/Lumeo.Editor.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
           dotnet pack src/Lumeo.FileViewer/Lumeo.FileViewer.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
           dotnet pack src/Lumeo.Scheduler/Lumeo.Scheduler.csproj -c Release --no-build -o ./nupkg /p:Version=${{ steps.ver.outputs.version }}
