+++ Lumeo.slnx
@@ -4,6 +4,7 @@
     <Project Path="src/Lumeo.Charts/Lumeo.Charts.csproj" />
     <Project Path="src/Lumeo.CodeEditor/Lumeo.CodeEditor.csproj" />
     <Project Path="src/Lumeo.DataGrid/Lumeo.DataGrid.csproj" />
+    <Project Path="src/Lumeo.DataGrid.Export/Lumeo.DataGrid.Export.csproj" />
     <Project Path="src/Lumeo.Editor/Lumeo.Editor.csproj" />
     <Project Path="src/Lumeo.FileViewer/Lumeo.FileViewer.csproj" />
     <Project Path="src/Lumeo.Gantt/Lumeo.Gantt.csproj" />
