+++ docs/Lumeo.Docs/Lumeo.Docs.csproj
@@ -94,6 +94,22 @@
     <BlazorWebAssemblyLazyLoad Include="Lumeo.Gantt.dll" />
     <BlazorWebAssemblyLazyLoad Include="Lumeo.PdfViewer.dll" />
     <BlazorWebAssemblyLazyLoad Include="Lumeo.Scheduler.dll" />
+
+    <!-- DataGrid Excel/PDF export backend + its heavy deps (~1.6 MB brotli). The DataGrid
+         itself is eager (Home hero), but export is an on-click action, so these load only
+         when a user actually exports — wired via DataGridExportLoader in Program.cs.
+         QuestPDF (PDF) throws on browser-wasm and is fetched only on a PDF click, so an
+         Excel-only WASM user never downloads it. -->
+    <BlazorWebAssemblyLazyLoad Include="Lumeo.DataGrid.Export.dll" />
+    <BlazorWebAssemblyLazyLoad Include="ClosedXML.dll" />
+    <BlazorWebAssemblyLazyLoad Include="ClosedXML.Parser.dll" />
+    <BlazorWebAssemblyLazyLoad Include="DocumentFormat.OpenXml.dll" />
+    <BlazorWebAssemblyLazyLoad Include="DocumentFormat.OpenXml.Framework.dll" />
+    <BlazorWebAssemblyLazyLoad Include="ExcelNumberFormat.dll" />
+    <BlazorWebAssemblyLazyLoad Include="RBush.dll" />
+    <BlazorWebAssemblyLazyLoad Include="SixLabors.Fonts.dll" />
+    <BlazorWebAssemblyLazyLoad Include="System.IO.Packaging.dll" />
+    <BlazorWebAssemblyLazyLoad Include="QuestPDF.dll" />
   </ItemGroup>
 
 </Project>
