+++ src/Lumeo/Lumeo.csproj
@@ -83,12 +83,13 @@
 
   <ItemGroup>
     <PackageReference Include="Blazicons.Lucide" Version="2.1.3" />
-    <PackageReference Include="ClosedXML" Version="0.104.2" />
     <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="10.0.6" />
     <!-- Source Link: embeds GitHub URLs in the PDB so debuggers can step into Lumeo source. -->
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
     <PackageReference Include="QRCoder" Version="1.8.0" />
-    <PackageReference Include="QuestPDF" Version="2024.12.3" />
+    <!-- ClosedXML (Excel) + QuestPDF (PDF) moved to the lazy-loaded Lumeo.DataGrid.Export
+         assembly so their ~1.6 MB of dependencies stay out of the core's eager first load.
+         The IDataGridExportService facade delegates Excel/PDF to that backend on demand. -->
   </ItemGroup>
 
   <!-- [LumeoForm] Roslyn source generator — builds a RenderFragment per annotated POCO. -->
